// src/core/http/generated/models/notificationTemplate.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */

export interface NotificationTemplate {
  id: string;
  key: string;
  name: string;
  /** @nullable */
  description?: string | null;
  module: string;
  event: string;
  channel: string;
  send_type: string;
  /** @nullable */
  subject?: string | null;
  body: string;
  variables?: string[];
  is_active: boolean;
  is_default: boolean;
  created_at?: string;
  updated_at?: string;
}