// src/core/http/generated/models/notificationDispatch.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationDispatchPayload } from './notificationDispatchPayload.ts';
import type { NotificationTemplate } from './notificationTemplate.ts';
import type { Tenant } from './tenant.ts';

export interface NotificationDispatch {
  id: string;
  tenant_id: string;
  notification_template_id: string;
  module: string;
  event: string;
  channel: string;
  recipient: string;
  contact: string;
  status: string;
  /** @nullable */
  subject?: string | null;
  body: string;
  /** @nullable */
  error_message?: string | null;
  payload?: NotificationDispatchPayload;
  created_at: string;
  /** @nullable */
  sent_at?: string | null;
  tenant?: Tenant;
  template?: NotificationTemplate;
}